requirements.txt
smolagents
python-telegram-bot